Refactor modelmeta module
